+++ tests/AvroConvertTests/FullSerializationAndDeserialization/AvroAttributeClassTests.cs
@@ -1,80 +1,74 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using AutoFixture;
+using FluentAssertions;
 using SolTechnology.Avro;
 using Xunit;
 
-namespace AvroConvertComponentTests.DefaultSerializationDeserialization
+namespace AvroConvertComponentTests.SerializationAndDeserialization
 {
     public class AvroAttributeClassTests
     {
-        private readonly Fixture _fixture;
+        private readonly Fixture _fixture = new();
 
-        public AvroAttributeClassTests()
-        {
-            _fixture = new Fixture();
-        }
-
-        [Fact]
-        public void Component_ClassContainsAvroAttributes_AttributeValuesAreResolved()
+        [Theory]
+        [MemberData(nameof(TestEngine.All), MemberType = typeof(TestEngine))]
+        public void Attribute_values_are_used_instead_of_property_names(Func<object, Type, dynamic> engine)
         {
             //Arrange
             AttributeClass toSerialize = _fixture.Create<AttributeClass>();
-
+            
 
             //Act
-            var result = AvroConvert.Serialize(toSerialize);
+            var deserialized = engine.Invoke(toSerialize, typeof(User));
 
-            var deserialized = AvroConvert.Deserialize<User>(result);
 
             //Assert
-            Assert.NotNull(result);
             Assert.NotNull(deserialized);
             Assert.Equal(toSerialize.NullableIntProperty, deserialized.favorite_number);
             Assert.Equal(toSerialize.NullableStringProperty, deserialized.name);
             Assert.Equal(toSerialize.AndAnotherString, deserialized.favorite_color);
         }
 
-        [Fact]
-        public void Component_FieldClassContainsAvroAttributes_AttributeValuesAreResolved()
+        [Theory]
+        [MemberData(nameof(TestEngine.All), MemberType = typeof(TestEngine))]
+        public void Attributes_are_used_instead_of_field_names(Func<object, Type, dynamic> engine)
         {
             //Arrange
             AttributeClassWithoutGetters toSerialize = _fixture.Create<AttributeClassWithoutGetters>();
 
 
             //Act
-            var result = AvroConvert.Serialize(toSerialize);
-
-            var deserialized = AvroConvert.Deserialize<User>(result);
+            var deserialized = engine.Invoke(toSerialize, typeof(User));
 
             //Assert
-            Assert.NotNull(result);
             Assert.NotNull(deserialized);
             Assert.Equal(toSerialize.NullableIntProperty, deserialized.favorite_number);
             Assert.Equal(toSerialize.StringProperty, deserialized.name);
             Assert.Equal(toSerialize.AndAnotherString, deserialized.favorite_color);
         }
 
-        [Fact]
-        public void Component_ComplexStruct_AttributeValuesAreResolved()
+        [Theory]
+        [MemberData(nameof(TestEngine.All), MemberType = typeof(TestEngine))]
+        public void Struct_containing_property(Func<object, Type, dynamic> engine)
         {
             //Arrange
             ComplexStruct toSerialize = new ComplexStruct(_fixture.Create<List<int>>());
 
 
             //Act
-            var result = AvroConvert.Serialize(toSerialize);
+            var deserialized = (ComplexStruct)engine.Invoke(toSerialize, typeof(ComplexStruct));
 
-            var deserialized = AvroConvert.Deserialize<ComplexStruct>(result);
 
             //Assert
-            Assert.NotNull(result);
-            Assert.Equal(toSerialize, deserialized);
+            deserialized.Should().BeEquivalentTo(toSerialize, opt => opt.ComparingByMembers<ComplexStruct>());
         }
 
-        [Fact]
-        public void Component_WriteSchemaContainsDefaultValuesForProperies_NullValuesAreReplacedWithDefault()
+        [Theory]
+        [MemberData(nameof(TestEngine.All), MemberType = typeof(TestEngine))]
+        public void Default_values_are_replacing_nulls(Func<object, Type, dynamic> engine)
         {
             //Arrange
             DefaultValueClass defaultValueClass = new DefaultValueClass();
@@ -83,12 +77,9 @@ namespace AvroConvertComponentTests.DefaultSerializationDeserialization
             defaultValueClass.andLongSmallDefaultedProperty = null;
 
             //Act
-            var result = AvroConvert.Serialize(defaultValueClass);
-
-            var deserialized = AvroConvert.Deserialize<DefaultValueClass>(result);
+            var deserialized = engine.Invoke(defaultValueClass, typeof(DefaultValueClass));
 
             //Assert
-            Assert.NotNull(result);
             Assert.NotNull(deserialized);
             Assert.Equal("Let's go", deserialized.justSomeProperty);
             Assert.Equal(9200000000000000007, deserialized.andLongBigDefaultedProperty);
@@ -96,19 +87,17 @@ namespace AvroConvertComponentTests.DefaultSerializationDeserialization
             Assert.Null(deserialized.andNullProperty);
         }
 
-        [Fact]
-        public void Component_ClassContainsAvroAttributes_IgnoredPropertiesAreNotSerialized()
+        [Theory]
+        [MemberData(nameof(TestEngine.All), MemberType = typeof(TestEngine))]
+        public void Properties_marked_as_ignored_are_not_serialized(Func<object, Type, dynamic> engine)
         {
             //Arrange
             AttributeClass toSerialize = _fixture.Create<AttributeClass>();
 
             //Act
-            var result = AvroConvert.Serialize(toSerialize);
-
-            var deserialized = AvroConvert.Deserialize<AttributeClass>(result);
+            var deserialized = engine.Invoke(toSerialize, typeof(AttributeClass));
 
             //Assert
-            Assert.NotNull(result);
             Assert.NotNull(deserialized);
             Assert.Null(deserialized.IgnoredProperty);
         }