+++ docs/CHANGELOG.md
@@ -331,3 +331,4 @@ START OF THE PROJECT
 - Changed GenerateClass default value representation
 - Extended GenerateModel to generalize complex type build of different types (ex. null, string, record) to object
 - Fix for GenerateModel to apply culture specific decimal separator
+- Avro2Json supports DefaultValue attribute